test: 테스트 메서드 분리
