+++ src/test/java/roomescape/MissionStepTest.java
@@ -16,7 +16,7 @@ import static org.hamcrest.Matchers.is;
 public class MissionStepTest {
 
     @Test
-    void 일단계() {
+    void 일단계_어드민_홈화면이_성공적으로_반환된다() {
         RestAssured.given().log().all()
                 .when().get("/admin")
                 .then().log().all()
@@ -24,7 +24,7 @@ public class MissionStepTest {
     }
 
     @Test
-    void 이단계() {
+    void 이단계_어드민_예약_페이지가_로딩된_후_예약_목록을_조회한다() {
         RestAssured.given().log().all()
                 .when().get("/admin/reservation")
                 .then().log().all()
@@ -38,7 +38,7 @@ public class MissionStepTest {
     }
 
     @Test
-    void 삼단계() {
+    void 삼단계_예약의_추가와_삭제가_이루어진다() {
         Map<String, String> params = new HashMap<>();
         params.put("name", "브라운");
         params.put("date", "2023-08-05");
@@ -68,27 +68,5 @@ public class MissionStepTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(0));
-    }
-
-    @Test
-    void 클라이언트_예약_생성_유효성_테스트() {
-        Map<String, String> params = new HashMap<>();
-        params.put("date", "2023-08-05");
-        params.put("time", "15:40");
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @Test
-    void 없는_인덱스로_예약을_삭제하는_경우() {
-        RestAssured.given().log().all()
-                .when().delete("/reservations/5")
-                .then().log().all()
-                .statusCode(404);
     }
 }
